refactor: update `fides init` to use the new documented config file and reorganize code locations
